requirements.txt
flask>=3.0
pypdf>=4.0
openpyxl>=3.1
pyyaml>=6.0
